+++ app/src/androidTest/java/com/supercilex/robotscouter/TmpTest.java
@@ -3,6 +3,8 @@ package com.supercilex.robotscouter;
 import android.support.test.rule.ActivityTestRule;
 import android.support.test.runner.AndroidJUnit4;
 
+import com.supercilex.robotscouter.ui.teamlist.TeamListActivity;
+
 import org.junit.Rule;
 import org.junit.Test;
 import org.junit.runner.RunWith;
@@ -13,7 +15,8 @@ import static android.support.test.espresso.Espresso.openActionBarOverflowOrOpti
 @RunWith(AndroidJUnit4.class)
 public class TmpTest {
     @Rule
-    public ActivityTestRule<MainActivity> mActivityTestRule = new ActivityTestRule<>(MainActivity.class);
+    public ActivityTestRule<TeamListActivity> mActivityTestRule = new ActivityTestRule<>(
+            TeamListActivity.class);
 
     @Test
     public void tmpTest() {
