--- app/src/main/java/com/supercilex/robotscouter/Utils.java
@@ -1,21 +1,0 @@
-package com.supercilex.robotscouter;
-
-import com.google.firebase.auth.FirebaseAuth;
-import com.google.firebase.auth.FirebaseUser;
-import com.google.firebase.database.FirebaseDatabase;
-
-public class Utils {
-    private static FirebaseDatabase sDatabase;
-
-    public static FirebaseDatabase getDatabase() {
-        if (sDatabase == null) {
-            sDatabase = FirebaseDatabase.getInstance();
-            sDatabase.setPersistenceEnabled(true);
-        }
-        return sDatabase;
-    }
-
-    public static FirebaseUser getUser() {
-        return FirebaseAuth.getInstance().getCurrentUser();
-    }
-}
