+++ app/src/main/java/com/supercilex/robotscouter/data/model/ScoutSpinner.java
@@ -1,0 +1,30 @@
+package com.supercilex.robotscouter.data.model;
+
+import com.google.firebase.database.DatabaseReference;
+
+import java.util.ArrayList;
+
+public class ScoutSpinner extends ScoutMetric<ArrayList<String>> {
+    private int mSelectedValue;
+
+    public ScoutSpinner() {
+    }
+
+    public ScoutSpinner(String name, ArrayList<String> values, int selectedValue) {
+        super(name, values);
+        mSelectedValue = selectedValue;
+    }
+
+    public int getSelectedValue() {
+        return mSelectedValue;
+    }
+
+    public void setSelectedValue(int selectedValue) {
+        mSelectedValue = selectedValue;
+    }
+
+    public void setSelectedValue(DatabaseReference databaseReference, int selectedValue) {
+        databaseReference.child("selectedValue").setValue(selectedValue);
+        mSelectedValue = selectedValue;
+    }
+}
