--- app/src/main/java/com/supercilex/robotscouter/models/scout/viewholders/SliderViewHolder.java
@@ -1,29 +1,0 @@
-package com.supercilex.robotscouter.models.scout.viewholders;
-
-import android.view.View;
-import android.widget.SeekBar;
-import android.widget.TextView;
-
-import com.google.firebase.database.DatabaseReference;
-import com.supercilex.robotscouter.R;
-import com.supercilex.robotscouter.models.scout.metrics.ScoutMetric;
-
-public class SliderViewHolder extends ScoutViewHolder {
-    private TextView mName;
-    private SeekBar mSeekBar;
-
-    public SliderViewHolder(View itemView) {
-        super(itemView);
-        mName = (TextView) itemView.findViewById(R.id.current_scout_text_view);
-        mSeekBar = (SeekBar) itemView.findViewById(R.id.seek_bar);
-    }
-
-    @Override
-    public void initialize(ScoutMetric view, DatabaseReference ref) {
-
-    }
-
-    public void setText(String name) {
-
-    }
-}
