+++ app/src/main/java/com/supercilex/robotscouter/ui/scout/viewholders/ScoutViewHolder.java
@@ -1,0 +1,15 @@
+package com.supercilex.robotscouter.ui.scout.viewholders;
+
+import android.support.v7.widget.RecyclerView;
+import android.view.View;
+
+import com.google.firebase.database.DatabaseReference;
+import com.supercilex.robotscouter.data.model.ScoutMetric;
+
+public abstract class ScoutViewHolder extends RecyclerView.ViewHolder {
+    public ScoutViewHolder(View itemView) {
+        super(itemView);
+    }
+
+    public abstract void initialize(ScoutMetric view, DatabaseReference ref);
+}
