+++ app/src/main/java/com/supercilex/robotscouter/ui/teamlist/NewTeamDialogFragment.java
@@ -1,0 +1,74 @@
+package com.supercilex.robotscouter.ui.teamlist;
+
+import android.annotation.SuppressLint;
+import android.app.Dialog;
+import android.content.DialogInterface;
+import android.os.Bundle;
+import android.support.annotation.NonNull;
+import android.support.v4.app.DialogFragment;
+import android.support.v7.app.AlertDialog;
+import android.text.TextUtils;
+import android.view.KeyEvent;
+import android.view.LayoutInflater;
+import android.view.View;
+import android.view.WindowManager;
+import android.view.inputmethod.EditorInfo;
+import android.widget.EditText;
+import android.widget.TextView;
+
+import com.supercilex.robotscouter.R;
+import com.supercilex.robotscouter.ui.scout.ScoutActivity;
+
+public class NewTeamDialogFragment extends DialogFragment {
+    @Override
+    public void onActivityCreated(Bundle savedInstanceState) {
+        super.onActivityCreated(savedInstanceState);
+        getDialog().getWindow()
+                .setSoftInputMode(WindowManager.LayoutParams.SOFT_INPUT_STATE_VISIBLE);
+    }
+
+    @NonNull
+    @Override
+    public Dialog onCreateDialog(Bundle savedInstanceState) {
+        AlertDialog.Builder builder = new AlertDialog.Builder(getActivity());
+        // Get the layout inflater
+        final LayoutInflater inflater = getActivity().getLayoutInflater();
+        @SuppressLint("InflateParams") final View view = inflater.inflate(R.layout.create_new_team_dialog,
+                                                                          null);
+        final EditText editText = (EditText) view.findViewById(R.id.create_new_team_dialog_team_number);
+
+        editText.setOnEditorActionListener(new TextView.OnEditorActionListener() {
+            public boolean onEditorAction(TextView v, int actionId, KeyEvent event) {
+                if ((event != null && (event.getKeyCode() == KeyEvent.KEYCODE_ENTER)) || (actionId == EditorInfo.IME_ACTION_DONE)) {
+                    if (!TextUtils.isEmpty(editText.getText())) {
+                        startScout(editText.getText().toString());
+                    }
+                }
+                return false;
+            }
+        });
+
+        // Inflate and set the layout for the dialog
+        // Pass null as the parent view because its going in the dialog layout
+        builder.setView(view)
+                // Add action buttons
+                .setPositiveButton(R.string.confirm, new DialogInterface.OnClickListener() {
+                    @Override
+                    public void onClick(DialogInterface dialog, int id) {
+                        if (!TextUtils.isEmpty(editText.getText())) {
+                            startScout(editText.getText().toString());
+                        } else {
+                            // TODO: 07/31/2016 cancel dismiss of dialog or show snackbar
+                        }
+                    }
+                }).setNegativeButton(R.string.cancel, new DialogInterface.OnClickListener() {
+            public void onClick(DialogInterface dialog, int id) {
+            }
+        });
+        return builder.create();
+    }
+
+    private void startScout(String teamNumber) {
+        startActivity(ScoutActivity.createIntent(getContext(), teamNumber, null));
+    }
+}
