+++ app/src/main/java/com/supercilex/robotscouter/util/Constants.java
@@ -1,0 +1,31 @@
+package com.supercilex.robotscouter.util;
+
+public class Constants {
+    public static final String TEAM_NICKNAME = "nickname";
+    public static final String TEAM_WEBSITE = "website";
+
+    public static final int CHECKBOX = 0;
+    public static final int COUNTER = 1;
+    public static final int SPINNER = 2;
+    public static final int EDIT_TEXT = 3;
+    public static final int SLIDER = 4;
+
+    public final static String INTENT_TEAM_NUMBER = "com.supercilex.robotscouter.TEAM_NUMBER";
+    public static final String INTENT_TEAM_KEY = "com.supercilex.robotscouter.TEAM_KEY";
+
+    public static final String FIREBASE_TEAM_INDEXES = "team-indexes";
+    public static final String FIREBASE_SCOUT_INDEXES = "scout-indexes";
+    public static final String FIREBASE_SCOUTS = "scouts";
+    public static final String FIREBASE_TEAMS = "teams";
+    public static final String FIREBASE_VIEWS = "views";
+
+    public static final String FIREBASE_TEAM_NAME = "name";
+    public static final String FIREBASE_TEAM_WEBSITE = "website";
+    public static final String FIREBASE_TEAM_MEDIA = "media";
+    public static final String FIREBASE_LAST_UPDATED = "lastUpdated";
+    public static final String FIREBASE_VALUE = "value";
+    public static final String FIREBASE_TYPE = "type";
+    public static final String FIREBASE_CUSTOM_NAME = "customName";
+    public static final String FIREBASE_CUSTOM_WEBSITE = "customWebsite";
+    public static final String FIREBASE_CUSTOM_MEDIA = "customMedia";
+}
