+++ app/src/main/java/com/supercilex/robotscouter/util/FirebaseUtils.java
@@ -1,0 +1,21 @@
+package com.supercilex.robotscouter.util;
+
+import com.google.firebase.auth.FirebaseAuth;
+import com.google.firebase.auth.FirebaseUser;
+import com.google.firebase.database.FirebaseDatabase;
+
+public class FirebaseUtils {
+    private static FirebaseDatabase sDatabase;
+
+    public static FirebaseDatabase getDatabase() {
+        if (sDatabase == null) {
+            sDatabase = FirebaseDatabase.getInstance();
+            sDatabase.setPersistenceEnabled(true);
+        }
+        return sDatabase;
+    }
+
+    public static FirebaseUser getUser() {
+        return FirebaseAuth.getInstance().getCurrentUser();
+    }
+}
