+++ app/src/main/java/com/supercilex/robotscouter/ztmpfirebase/FirebaseArray.java
@@ -1,0 +1,129 @@
+package com.supercilex.robotscouter.ztmpfirebase;
+
+/*
+ * Copyright 2016 Google Inc. All Rights Reserved.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except
+ * in compliance with the License. You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software distributed under the
+ * License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either
+ * express or implied. See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+import com.google.firebase.database.ChildEventListener;
+import com.google.firebase.database.DataSnapshot;
+import com.google.firebase.database.DatabaseError;
+import com.google.firebase.database.Query;
+
+import java.util.ArrayList;
+import java.util.List;
+
+/**
+ * This class implements an array-like collection on top of a Firebase location.
+ */
+class FirebaseArray implements ChildEventListener {
+    private Query mQuery;
+    private OnChangedListener mListener;
+    private List<DataSnapshot> mSnapshots = new ArrayList<>();
+
+    public FirebaseArray(Query ref) {
+        mQuery = ref;
+        mQuery.addChildEventListener(this);
+    }
+
+    public void cleanup() {
+        mQuery.removeEventListener(this);
+    }
+
+    public int getCount() {
+        return mSnapshots.size();
+    }
+
+    public DataSnapshot getItem(int index) {
+        return mSnapshots.get(index);
+    }
+
+    private int getIndexForKey(String key) {
+        int index = 0;
+        for (DataSnapshot snapshot : mSnapshots) {
+            if (snapshot.getKey().equals(key)) {
+                return index;
+            } else {
+                index++;
+            }
+        }
+        throw new IllegalArgumentException("Key not found");
+    }
+
+    @Override
+    public void onChildAdded(DataSnapshot snapshot, String previousChildKey) {
+        int index = 0;
+        if (previousChildKey != null) {
+            index = getIndexForKey(previousChildKey) + 1;
+        }
+        mSnapshots.add(index, snapshot);
+        notifyChangedListeners(OnChangedListener.EventType.Added, index);
+    }
+
+    @Override
+    public void onChildChanged(DataSnapshot snapshot, String previousChildKey) {
+        int index = getIndexForKey(snapshot.getKey());
+        mSnapshots.set(index, snapshot);
+        notifyChangedListeners(OnChangedListener.EventType.Changed, index);
+    }
+
+    @Override
+    public void onChildRemoved(DataSnapshot snapshot) {
+        int index = getIndexForKey(snapshot.getKey());
+        mSnapshots.remove(index);
+        notifyChangedListeners(OnChangedListener.EventType.Removed, index);
+    }
+
+    @Override
+    public void onChildMoved(DataSnapshot snapshot, String previousChildKey) {
+        int oldIndex = getIndexForKey(snapshot.getKey());
+        mSnapshots.remove(oldIndex);
+        int newIndex = previousChildKey == null ? 0 : getIndexForKey(previousChildKey) + 1;
+        mSnapshots.add(newIndex, snapshot);
+        notifyChangedListeners(OnChangedListener.EventType.Moved, newIndex, oldIndex);
+    }
+
+    @Override
+    public void onCancelled(DatabaseError databaseError) {
+        notifyCancelledListeners(databaseError);
+    }
+
+    public void setOnChangedListener(OnChangedListener listener) {
+        mListener = listener;
+    }
+
+    protected void notifyChangedListeners(OnChangedListener.EventType type, int index) {
+        notifyChangedListeners(type, index, -1);
+    }
+
+    protected void notifyChangedListeners(OnChangedListener.EventType type,
+                                          int index,
+                                          int oldIndex) {
+        if (mListener != null) {
+            mListener.onChanged(type, index, oldIndex);
+        }
+    }
+
+    protected void notifyCancelledListeners(DatabaseError databaseError) {
+        if (mListener != null) {
+            mListener.onCancelled(databaseError);
+        }
+    }
+
+    public interface OnChangedListener {
+        void onChanged(EventType type, int index, int oldIndex);
+
+        void onCancelled(DatabaseError databaseError);
+
+        enum EventType {Added, Changed, Removed, Moved}
+    }
+}
