+++ app/src/main/res/layout/activity_main.xml
@@ -7,7 +7,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:fitsSystemWindows="true"
-    tools:mContext="com.supercilex.robotscouter.MainActivity">
+    tools:mContext="com.supercilex.robotscouter.ui.teamlist.TeamListActivity">
 
     <android.support.design.widget.AppBarLayout
         android:layout_width="match_parent"
