+++ gradle.properties
@@ -1,4 +1,4 @@
-#Sat Sep 24 19:47:46 PDT 2016
+#Sun Oct 02 14:35:00 PDT 2016
 org.gradle.jvmargs=-Xmx2048m
-VERSION_CODE=1034
+VERSION_CODE=1037
 android.enableBuildCache=true
